requirements.txt
certifi==2018.8.24
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.1
django-allauth==0.37.1
django-daterange-filter==1.3.0
django-filter==2.0.0
django-import-export==1.0.1
django-js-asset==1.0.0
django-mptt==0.9.0
django-webpack-loader==0.6.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.7
jdcal==1.4
Markdown==2.6.11
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.5.7
Pillow==5.0.0
pkg-resources==0.0.0
python3-openid==3.1.0
pytz==2017.3
PyYAML==3.13
requests==2.19.1
requests-oauthlib==1.0.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.23
xlrd==1.1.0
xlwt==1.3.0
yolk==0.4.3
